[Custom ORS Preparation] Add dependencies to graphhopper-reader-osm
Declared dependencies at the pom.xml of the sub-project
graphhopper-reader-osm:

* com.graphhopper, graphhopper-api
* com.carrotsearch, hppc

Currently, the packaging of the sub-projects is done by hand,
one by one, invoking maven directly in the command line.
This leads to dependencies not being package in the JARs
being created.

Also, the sub-project graphhopper-reader-osm depends on the
graphhopper-core sub-project. graphhopper-reader-osm would
access the dependencies above throught graphhopper-core.
Because the packaging is done by hand, one must declare all the
dependencies that graphhopper-reader-osm needs.